Using the Cached Container
Copy the path to the file and create a new  variable way up top before we even start with the ContainerBuilder. Our app now has two options: we can create the ContainerBuilder, load it up with the Definition config and then use it, OR, if that cached container is available, we can skip everything and just use it. After all. if we call get('logger') on it, it'll give us the exact same Logger.

But one way or another, that file eventually exists. So if we require it, we can say {dollar_sign}container = new \ProjectServiceContainer(), which is the class name used in the cache file.

We're still passing this {dollar_sign}container into runApp(), and even though it's technically a different object, it's not going to make any difference. The only thing we need to change is that runApp() is type-hinted with ContainerBuilder. Well, it turns out that what we really need is Container, which is the base class for the builder and our cached class.

So I'll change the type-hint to Container. And we can go a step further: the Container class implements an interface called ContainerInterface.

Ok, try out the brand new cached container!

php dino_container/roar.php
It works! And woh - check out that elapsed time: 4ms, down from 19. If you delete the cached_container.php file, the next run takes 22ms because it needs to rebuild it. Then we're right back down to 4ms. This is one reason why Symfony is able to be so fast, even in big systems.

Now that you've got the real story of how container building works, let's see how things look inside Symfony.

+++ dino_container/roar.php
@@ -2,7 +2,7 @@
 
 namespace Dino\Play;
 
-use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\Dumper\PhpDumper;
@@ -11,23 +11,29 @@ require __DIR__.'/../vendor/autoload.php';
 
 $start = microtime(true);
 
-$container = new ContainerBuilder();
-$container->setParameter('root_dir', __DIR__);
+$cachedContainer = __DIR__.'/cached_container.php';
+if (!file_exists($cachedContainer)){
+  $container = new ContainerBuilder();
+  $container->setParameter('root_dir', __DIR__);
 
-$loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/config'));
-$loader->load('services.yml');
+  $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/config'));
+  $loader->load('services.yml');
 
-$container->compile();
+  $container->compile();
 
-$dumper = new PhpDumper($container);
-file_put_contents(__DIR__.'/cached_container.php', $dumper->dump());
+  $dumper = new PhpDumper($container);
+  file_put_contents(__DIR__.'/cached_container.php', $dumper->dump());
+}
+
+require $cachedContainer;
+$container = new \ProjectServiceContainer();
 
 runApp($container);
 
 $elapsed = round((microtime(true) - $start) * 1000);
 $container->get('logger')->debug('Elapsed Time: '.$elapsed.'ms');
 
-function runApp(ContainerBuilder $container){
+function runApp(ContainerInterface $container){
   $container->get('logger')->info('ROOOOAR'); 
 }
